Moved the classes into application specific package.
This enables us to separate the application logic with the cross-cutting logic.
