+++ src/main/kotlin/com/test/product/application/repositories/ProductRepository.kt
@@ -1,6 +1,6 @@
-package com.test.product.repositories
+package com.test.product.application.repositories
 
-import com.test.product.domain.Product
+import com.test.product.application.domain.Product
 import org.springframework.data.repository.CrudRepository
