+++ src/main/kotlin/com/test/product/application/service/ProductService.kt
@@ -1,7 +1,10 @@
-package com.test.product.service
+package com.test.product.application.service
 
-import com.test.product.domain.Product
-import com.test.product.repositories.ProductRepository
+import com.test.product.application.domain.Product
+import com.test.product.application.repositories.ProductRepository
+import com.test.product.cross_cutting.ProductRepositoryMetrics
+import com.test.product.cross_cutting.ProductRepositoryMetricsAdvice
+import com.test.product.cross_cutting.ProductRepositoryMetricsAdvice.*
 import org.springframework.stereotype.Service
 import java.util.*
 
@@ -9,10 +12,12 @@ import java.util.*
 @Service
 class ProductService(private val productRepository: ProductRepository){
 
+    @ProductRepositoryMetrics
     fun getProducts(): List<Product> {
         return productRepository.findAll().toList();
     }
 
+    @ProductRepositoryMetrics
     fun getProductById(id: String): Optional<Product> {
         if (!isValidId(id)) return Optional.empty();
         return productRepository.findById(id);
